requirements.txt
Django==2.2.9
django-filter==2.3.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
